BK-1664 - Update script to fix issue with menu when viewed in Editor.
The event handling script was updating the CSS class for the body tag rather
then appending a class to it.

I have replaced the script the a jQuery'd version as we're including the
framework anyway and it takes care of other issues with the way the event was
bound to the element.